requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-filter==23.5
django-storages==1.14.2
firebase-admin==6.4.0
requests==2.31.0
python-jose==3.3.0
django-rest-auth==0.9.5
django-allauth==0.61.1 